mockup

+++ lib/BBC/Commands/cpan_river.pm
@@ -7,6 +7,8 @@ use warnings;
 use Test::More;    # for debugging
 use Mojo::Base 'Mojolicious::Command';
 
+use Getopt::Long qw(GetOptionsFromArray);
+
 # Short description
 has description => 'update cpan river list';
 
@@ -19,26 +21,83 @@ has dry_run => sub { 0 };
 sub run {
     my ( $self, @args ) = @_;
 
-    my $dry_run = grep { $_ =~ qr{dry-run} } @args;
-    my $force   = grep { $_ =~ qr{force} } @args;
+    my ( $dry_run, $force, $csv_file, $limit, $debug );
+    GetOptionsFromArray( 
+        \@args, 
+        # 'dry-run', \$dry_run, # not implemented yet
+        'force', \$force,
+        'debug', \$debug,
+        'csv=s', \$csv_file,
+        'limit=i', \$limit,
+    );
 
-    my $backend = $self->app->bbc->backend;
+    die qq[Missing --csv argument\n] unless $csv_file;
+    die qq[Cannot find csv file $csv_file\n] unless -e $csv_file;
     
     $self->dry_run($dry_run);
     note "dry-run mode enable" if $self->dry_run;
 
+    my $backend = $self->app->bbc->backend;
     die qq[Cannot ping db\n] unless $backend->ping;
 
     note "Deleting tables...";
+
     $backend->delete( 'cpan_river' );
     $backend->delete( 'maintainers' );
     $backend->delete( 'distro_maintainers' );
     
-    note "Inserting new values";
+    my $cache_authors = {}; # cache for authors
 
-    $backend->insert('cpan_river', { distribution => 'test', state => "xxx", counter => 1234, updated_at => \"now()" } );
+    note qq[Reading CSV file];
+    open( my $fh, '<', $csv_file ) or die $!;
+    my $c = 0;
+    while ( my $line = readline($fh) ) {
+        ++$c;
+        note $line if $debug;
+        my @row = split(/\s*,\s*/, $line );
 
-    note explain $backend->select( 'cpan_river' )->hash;
+        if ( $c == 1 ) {
+            die "Invalid csv format: ", explain \@row unless $row[0] eq 'count'
+                && $row[1] eq 'distribution'
+                && $row[2] eq 'core_upstream_status'
+                && $row[3] eq 'maintainers';
+            next;
+        }
+        
+        my ( $count, $distribution, $state, $maintainers_str ) = @row;
+        $maintainers_str //= '';
+        $state //= '';
+        next unless defined $distribution;
+
+        # insert the distro
+        $backend->insert('cpan_river', 
+            { distribution => $distribution, state => $state, counter => $count, updated_at => \"now()" } 
+        );
+        my $dist_id = $backend->select( 'cpan_river', 'id', { distribution => $distribution }, { limit => 1 } )->array->[0];
+
+        my @maintainers = split( /\s+/, $maintainers_str );
+        foreach my $maint ( @maintainers ) {
+            if ( ! $cache_authors->{$maint} ) { 
+                $backend->insert('maintainers', { name => $maint }, {on_conflict => undef} );            
+                $cache_authors->{$maint} = $backend->select( 'maintainers', 'id', { name => $maint }, { limit => 1 } )->array->[0];
+            }
+            $backend->insert('distro_maintainers', 
+                { distro_id => $dist_id, maintainer_id => $cache_authors->{$maint} } 
+            );
+        }
+
+        note explain [ $count, $distribution, $state, $maintainers_str ] if $debug;
+        last if $limit && $c > $limit;
+    }
+
+    note qq[Mark the top 1000];
+
+    $backend->pg->db->query( "update cpan_river set is_top_1000=true 
+        where id in
+        ( select id from cpan_river order by counter desc limit 1000 )
+        ");
+
+    note qq[Done.\n];
 
     return;
 }
@@ -47,10 +106,12 @@ sub run {
 
 =head1 SYNOPSIS
 
-  Usage: APPLICATION cpan_river file.csv
+  Usage: APPLICATION cpan_river --csv=file.csv
+
+Clear and restore database for cpan river.
 
   Options:
 
-    --dry-run to do not unlink the binaries
+    --csv=FILE  load CSV file
 
 =cut
